Additional Support for ATSC
Ther ATSC tuners in Enigma2 are not usefull at all because some parts were missing.

This commit allows ATSC tuners to be configured, parses atsc.xml file, reads lamedb and finally scans for channels.

There are more parts missing, so consider this as work in progress.

Hopefully with reports and logs from people with ATSC tuner and reception will move on...